Renomeamento do pacote de domain entities para bookloc.
