+++ grails-app/domain/bookloc/Book.groovy
@@ -1,0 +1,18 @@
+package bookloc
+
+class Book {
+    String title
+    String publisher
+    String isbn
+    Boolean isRented
+    Long shelf
+    Long rack
+
+    static constraints = {
+        isRented nullable: true
+    }
+    static mapping = {
+        isRented defaultValue: "false"
+    }
+}
+
