+++ grails-app/domain/bookloc/User.groovy
@@ -1,0 +1,16 @@
+package bookloc
+
+class User {
+    String name
+    String password
+    String cpf
+    Date dateOfBirth
+    String rentedBooks
+
+    static constraints = {
+        rentedBooks nullable: true
+    }
+    static mapping = {
+        rentedBooks defaultValue: "''"
+    }
+}
