--- grails-app/domain/entidades/Librarian.groovy
@@ -1,11 +1,0 @@
-package entidades
-
-class Librarian {
-    String name
-    String password
-    String cpf
-    Date dateOfBirth
-
-    static constraints = {
-    }
-}
